feat:直接从服务端进行CURD操作

+++ apps/main-app/utils/actions.ts
@@ -8,7 +8,7 @@ import { db } from '~/db';
 import { task } from '~/db/schema';
 
 interface FormData {
-  get(key: string): string;
+  get(key: string): FormDataEntryValue | null;
 }
 interface PrevState {
   message: string;
@@ -19,8 +19,10 @@ export const getAllTasks = async () => {
 };
 
 export const createTask = async (formData: FormData) => {
-  const content: string = formData.get('content');
-
+  const content = formData.get('content');
+  if (typeof content !== 'string') {
+    return;
+  }
   await db.insert(task).values({
     id: createId(),
     content: content,
@@ -28,16 +30,18 @@ export const createTask = async (formData: FormData) => {
   revalidatePath('/tasks');
 };
 
-export const createTaskCustom = async (
-  prevState: PrevState,
-  formData: FormData
-) => {
+export const createTaskCustom = async (prevState: void, formData: FormData) => {
   // await new Promise((resolve) => setTimeout(resolve, 2000));
 
   const content = formData.get('content');
   const Task = z.object({
     content: z.string().min(5),
   });
+
+  if (typeof content !== 'string') {
+    return;
+  }
+
   try {
     Task.parse({ content });
     await db.insert(task).values({
@@ -52,20 +56,35 @@ export const createTaskCustom = async (
 
 export const deleteTask = async (formData: FormData) => {
   const id = formData.get('id');
-
+  if (typeof id !== 'string' || id.length === 0) {
+    return;
+  }
   await db.delete(task).where(eq(task.id, id));
   revalidatePath('/tasks');
 };
 
 export const getTaskById = async (id: string) => {
-  return db.select().from(task).where(eq(task.id, id));
+  const data = await db
+    .select({
+      id: task.id,
+      content: task.content,
+      completed: task.completed,
+      createdAt: task.createdAt,
+    })
+    .from(task)
+    .where(eq(task.id, id))
+    .limit(1);
+
+  return data[0];
 };
 
 export const editTask = async (formData: FormData) => {
   const id = formData.get('id');
   const content = formData.get('content');
   const completed = formData.get('completed');
-
+  if (typeof id !== 'string' || typeof content !== 'string') {
+    return;
+  }
   await db
     .update(task)
     .set({
